requirements.txt
numpy
tensorflow
keras
nltk
pyyaml
fastapi
pydantic
uvicorn